refactor: move volume methods from main class into separate mixin
